i made a booboo

+++ src/MAndApps/apps/spacewars/entity/enemy/GreenEnemy.java
@@ -142,7 +142,7 @@ public class GreenEnemy extends Enemy {
         g.setColor(color);
         int temp;
         try{
-            temp = r.nextInt((int)(0-((double)absoluteTime/20d))+5);
+            temp = r.nextInt((int)(0-(absoluteTime/20d))+5);
         }catch(Exception e){
             temp = 0;
         }
